Add `ag` (The Silver Searcher) to the eng build
Problem
-------

Neither `ack` nor `ag` are installed by default.  These tools both
provide `grep` like functionality but are tailored a bit for searching
through code.  `ag` was developed as an `ack` replacement and was built
for speed.

Solution
--------

Add `ag` as a default package to install with the rest of the eng stuff.

Readings
--------

https://github.com/ggreer/the_silver_searcher